+++ src/taskpane/excel-office-document.js
@@ -1,0 +1,18 @@
+/* global Excel console */
+
+const insertExcelText = async (text) => {
+    // Write text to the top left cell.
+    try {
+      Excel.run(async (context) => {
+        const sheet = context.workbook.worksheets.getActiveWorksheet();
+        const range = sheet.getRange("A1");
+        range.values = [[text]];
+        range.format.autofitColumns();
+        return context.sync();
+      });
+    } catch (error) {
+      console.log("Error: " + error);
+    }
+  };
+  
+  export default insertExcelText;
