+++ src/taskpane/word-office-document.js
@@ -1,0 +1,16 @@
+/* global Word console */
+
+const insertWordText = async (text) => {
+    // Write text to the document.
+    try {
+      await Word.run(async (context) => {
+        let body = context.document.body;
+        body.insertParagraph(text, Word.InsertLocation.end);
+        await context.sync();
+      });
+    } catch (error) {
+      console.log("Error: " + error);
+    }
+  };
+  
+  export default insertWordText;
